+++ src/nextjs/templates.ts
@@ -149,7 +149,9 @@ export function getSentryConfigContents(
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/`;
   }
 
-  const integrationsOptions = getClientIntegrationsSnippet(selectedFeaturesMap);
+  const integrationsOptions = getClientIntegrationsSnippet({
+    replay: config === 'client' && selectedFeaturesMap.replay,
+  });
 
   let replayOptions = '';
   if (config === 'client') {
@@ -367,7 +369,9 @@ YourCustomErrorComponent.getInitialProps = async (contextData${
 export function getInstrumentationHookContent(
   instrumentationHookLocation: 'src' | 'root',
 ) {
-  return `export async function register() {
+  return `import * as Sentry from '@sentry/nextjs';
+
+export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
     await import('${
       instrumentationHookLocation === 'root' ? '.' : '..'
@@ -380,6 +384,8 @@ export function getInstrumentationHookContent(
     }/sentry.edge.config');
   }
 }
+
+export const onRequestError = Sentry.captureRequestError;
 `;
 }
 
@@ -387,7 +393,9 @@ export function getInstrumentationHookCopyPasteSnippet(
   instrumentationHookLocation: 'src' | 'root',
 ) {
   return makeCodeSnippet(true, (unchanged, plus) => {
-    return unchanged(`export ${plus('async')} function register() {
+    return unchanged(`${plus("import * as Sentry from '@sentry/nextjs';")}
+
+export ${plus('async')} function register() {
   ${plus(`if (process.env.NEXT_RUNTIME === 'nodejs') {
     await import('${
       instrumentationHookLocation === 'root' ? '.' : '..'
@@ -399,7 +407,10 @@ export function getInstrumentationHookCopyPasteSnippet(
       instrumentationHookLocation === 'root' ? '.' : '..'
     }/sentry.edge.config');
   }`)}
-}`);
+}
+
+${plus('export const onRequestError = Sentry.captureRequestError;')}
+`);
   });
 }